requirements.txt
Flask>=2.0.0
Flask-JWT-Extended>=4.4.0
Flask-Cors>=3.0.10
mysql-connector-python>=8.0.30
Werkzeug>=2.0.0
